consolidate withProperties test into main useOsdkObjects test file
